refactor(build): 의존관계 수정으로 clean architecture 구조로 변경
- 모듈 간의 의존관계
    - presentation -> application <- infrastructure (큰 흐름)
    - presentation -> domain : application의 응답을 도메인 모델로 받기 위함
    - application -> domain : application 에서는 프로그램의 흐름을 제어하는 서비스로직 담당, 도메인의 로직(도메인 내부 메서드), 도메인 간의 상호 로직(추 후, domain service로 명명)은 해당 모듈에서 가지고 오도록 하여, applicaiton은 오로지 프로그램의 흐름을 제어하는 로직이 들어가도록 설정
    - infrastructure -> domain : applicaton 모듈의 요청을 특정기술의 모델(ex> jpaEntity)이 아닌 domain 엔티티로 전달하기 위함
    - presentation -> infrastructure : 단순히 infrastructure의 빈들을 application에서도 사용가능하도록 등록하기 위함.

- 해당 구조에 맞게 클래스 파일의 이동
