+++ module-application/build.gradle
@@ -20,7 +20,6 @@ version = '0.0.1-SNAPSHOT'
 
 dependencies {
     implementation project(':module-domain')
-    implementation project(':module-infrastructure')
     implementation 'org.springframework.boot:spring-boot-starter'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
