+++ module-application/src/main/java/project/redis/application/cinema/port/outbound/CinemaQueryPort.java
@@ -1,4 +1,4 @@
-package project.redis.infrastructure.cinema.inbound.port;
+package project.redis.application.cinema.port.outbound;
 
 import java.util.List;
 import project.redis.domain.cinema.Cinema;
