+++ module-application/src/main/java/project/redis/application/genre/port/inbound/GenreQueryUseCase.java
@@ -1,0 +1,9 @@
+package project.redis.application.genre.port.inbound;
+
+import java.util.List;
+import project.redis.domain.genre.Genre;
+
+public interface GenreQueryUseCase {
+
+    List<Genre> getGenres();
+}
