+++ module-application/src/main/java/project/redis/application/genre/port/outbound/GenreQueryPort.java
@@ -1,0 +1,9 @@
+package project.redis.application.genre.port.outbound;
+
+import java.util.List;
+import project.redis.domain.genre.Genre;
+
+public interface GenreQueryPort {
+
+    List<Genre> findAllGenres();
+}
