+++ module-application/src/main/java/project/redis/application/screening/service/ScreeningQueryService.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import project.redis.application.screening.port.inbound.ScreeningQueryUseCase;
 import project.redis.application.screening.port.inbound.ScreeningsQueryParam;
+import project.redis.application.screening.port.outbound.ScreeningQueryPort;
 import project.redis.domain.screening.Screening;
-import project.redis.infrastructure.screening.inbound.port.inbound.ScreeningQueryPort;
 
 
 
@@ -14,7 +14,9 @@ import project.redis.infrastructure.screening.inbound.port.inbound.ScreeningQuer
 @RequiredArgsConstructor
 public class ScreeningQueryService implements ScreeningQueryUseCase {
 
+
     private final ScreeningQueryPort screeningQueryPort;
+
 
     @Override
     public List<Screening> getScreenings(ScreeningsQueryParam param) {
