+++ module-infrastructure/build.gradle
@@ -20,6 +20,7 @@ version = '0.0.1-SNAPSHOT'
 
 dependencies {
     implementation project(':module-domain')
+    implementation project(':module-application')
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     runtimeOnly 'com.mysql:mysql-connector-j'
     implementation 'org.flywaydb:flyway-core'
