+++ module-presentation/build.gradle
@@ -20,6 +20,7 @@ version = '0.0.1-SNAPSHOT'
 
 dependencies {
     implementation project(':module-application')
+    implementation project(':module-infrastructure')
     implementation project(':module-domain')
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
