+++ module-presentation/src/main/java/project/redis/presentation/TheaterApplication.java
@@ -8,9 +8,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication(scanBasePackages = {
         "project.redis.application",
         "project.redis.presentation",
+        "project.redis.domain",
         "project.redis.infrastructure"
 })
-
 public class TheaterApplication {
     public static void main(String[] args) {
         SpringApplication.run(TheaterApplication.class, args);
